Fix: Move 12-2 to ./campaign/campaign_main
